Defined 24km EASE2 grids in relation to SMAP 03km.
For CETB SMAP processing, we need a base grid resolution that is a
power-of-2 multiple of the N/S/M03km grids. We Cannot change the 03km
grid definitions, as they are fixed and adopted already by the SMAP
project and projects.  So these 24km grids are exact multiples of 8 of
the respective 03 km grids.